risk controller attempt
modified environment added SOCP risk controller.
RL and controller need a lot of fine-tuning, as of now, the risk optimizer leads to net 0 profit or loss, which is technically an improvement from running in loss lol.
but, I feel like it isn't doing its job well, or the constraints are too tight. 
params need tuning.